java/src/test/java/im/wangbo/java/leetcode/string/LongestPalindromeSolutionTest.java
package im.wangbo.java.leetcode.string;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

/**
 * TODO add brief description here
 *
 * @author Elvis Wang
 */
class LongestPalindromeSolutionTest {

    @ParameterizedTest
    @MethodSource("arguments")
    void test(final String str, final String expected) {
        final LongestPalindromeSolution s = new LongestPalindromeSolution();
        final String result = s.longestPalindrome(str);

        Assertions.assertThat(result).isEqualTo(expected);
    }

    static Stream<Arguments> arguments() {
        return Stream.of(
                Arguments.of("babad", "bab"),
                Arguments.of("cbbd", "bb")
        );
    }
}